Preventing game starting in DMChannel and moved State object
