--- src/utils/structures/NisteryStates.js
@@ -1,5 +1,0 @@
-module.exports = {
-    INTRO: 1,
-    GAME: 2,
-    INVALID: 3
-}
